+++ test/ad_size_test.dart
@@ -39,5 +39,11 @@ void main() {
       () => const AdSize.inline().suggestedHeightDp,
       throwsA(isA<StateError>()),
     );
+    expect(
+      () => const AdSize.collapsible(
+        placement: CollapsiblePlacement.bottom,
+      ).suggestedHeightDp,
+      throwsA(isA<StateError>()),
+    );
   });
 }